# api-docs/openapi.yaml
openapi: 3.1.0
info:
  title: Advanced Rust Payroll System API
  description: API for a high-performance, next-generation payroll management system.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /roles:
    post:
      summary: Create a New Role
      description: Creates a new role configuration.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleConfiguration'
      responses:
        '201':
          description: Role created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleConfiguration'
    get:
      summary: List All Roles
      description: Retrieves a list of all role configurations.
      responses:
        '200':
          description: A list of role configurations.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoleConfiguration'
  /roles/{role_id}:
    put:
      summary: Update a Role
      description: Updates an existing role configuration.
      parameters:
        - name: role_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleConfiguration'
      responses:
        '200':
          description: Role updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleConfiguration'
    get:
      summary: Get Role by ID
      description: Retrieves a role configuration by its ID.
      parameters:
        - name: role_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleConfiguration'
    delete:
      summary: Delete a Role
      description: Deletes a role configuration by its ID.
      parameters:
        - name: role_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Role deleted successfully.
...
components:
  schemas:
...
    RoleConfiguration:
      type: object
      required:
        - role_id
        - role_name
        - base_monthly_salary
        - currency
        - overtime_rate_multiplier
        - leave_policies
        - working_hours_per_day
        - is_active
      properties:
        role_id:
          type: string
        role_name:
          type: string
        base_monthly_salary:
          type: number
          format: double
        currency:
          type: string
        overtime_rate_multiplier:
          type: number
          format: double
        leave_policies:
          type: array
          items:
            $ref: '#/components/schemas/LeavePolicy'
        working_hours_per_day:
          type: number
          format: double
        is_active:
          type: boolean
    LeavePolicy:
      type: object
      required:
        - leave_type_id
        - salary_deduction_percent
        - requires_approval
        - carry_forward_allowed
      properties:
        leave_type_id:
          $ref: '#/components/schemas/LeaveTypeId'
        salary_deduction_percent:
          type: number
          format: double
        max_days_per_month:
          type: integer
          format: int32
          nullable: true
        max_days_per_year:
          type: integer
          format: int32
          nullable: true
        requires_approval:
          type: boolean
        carry_forward_allowed:
          type: boolean
    LeaveTypeId:
      type: string
...